backend/requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
colorama==0.4.4
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-summernote==0.8.11.6
django-taggit==1.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pydot==1.4.2
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
